requirements.txt
torch==2.3.0
torchvision==0.18.0
pillow==10.3.0
matplotlib==3.8.4
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9